refactor: 코드 스플릿 완료
refactor: 코드 스플릿 완료
- 컬러 변경
