+++ src/components/Cursor/Cursor.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function Cursor() {
+  return (
+    <span className="animate-blink absolute top-0.5 left-0 h-6 w-0.5 bg-black"></span>
+  );
+}
